Disable windmill / docker jobs
We need to disable windmill jobs, as we are currently wedged on fixing
jobs. Current fixes need to happen our a few projets, so once everythign
is working, we'll bring this back online and merge eveyrthing in at
once.

Docker jobs look to be broken, disable for now and maybe look to migrate
to podman.

Add pytest to test-requirements.txt

Change-Id: Ie7b89f4bf0089e8020a625d4c561089695d529c0
Signed-off-by: Paul Belanger <pabelanger@redhat.com>